SXDEDPCXZIC-24_DATAVIC-471 / simplify actions
