+++ README.md
@@ -1,21 +1,6 @@
 # ckanext-datavic-reporting
+
 Custom CKAN reporting extension for Data.Vic
-
-### Create Database Tables
-
-This extension utilises two new database tables in the main CKAN database:
-
-__report_schedule:__ stores scheduled report configuration settings.
-
-__report:__ an instance of a report generate via the scheduled report.
-
-Use the following paster command to initialise the tables:
-
-    . /app/ckan/default/bin/activate
-
-    ckan datavic_reporting initdb --config=/app/ckan/default/production.ini
-
-* Replacing `production.ini` with the respective CKAN .ini file to be used.
 
 ### Create scheduled reporting frequencies
 
@@ -48,7 +33,7 @@ __verb:__ POST
 
 __params:__
 
-    report_type - 
+    report_type -
     org_id - organisation UUID
     sub_org_ids - comma separated list of sub-org UUIDs
     frequency - 'monthly' or 'yearly'
@@ -62,7 +47,7 @@ __verb:__ POST
 __params:__
 
     id - report_schedule.id record to be deleted
-    report_type - 
+    report_type -
     org_id - organisation UUID
     sub_org_ids - comma separated list of sub-org UUIDs
     frequency - 'monthly' or 'yearly'
