--- ckanext/datavic_reporting/logic/action/__init__.py
@@ -1,11 +1,0 @@
-# encoding: utf-8
-
-# this is a namespace package
-try:
-    import pkg_resources
-
-    pkg_resources.declare_namespace(__name__)
-except ImportError:
-    import pkgutil
-
-    __path__ = pkgutil.extend_path(__path__, __name__)
