+++ ckanext/datavic_reporting/migration/datavic_reporting/env.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
 
 from __future__ import with_statement
-from alembic import context
-from sqlalchemy import engine_from_config, pool
-from logging.config import fileConfig
 
 import os
+from logging.config import fileConfig
+
+from alembic import context
+from sqlalchemy import engine_from_config, pool
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
@@ -42,10 +43,12 @@ def run_migrations_offline():
 
     """
 
-    url = config.get_main_option(u"sqlalchemy.url")
+    url = config.get_main_option("sqlalchemy.url")
     context.configure(
-        url=url, target_metadata=target_metadata, literal_binds=True,
-        version_table=u'{}_alembic_version'.format(name)
+        url=url,
+        target_metadata=target_metadata,
+        literal_binds=True,
+        version_table="{}_alembic_version".format(name),
     )
 
     with context.begin_transaction():
@@ -61,14 +64,15 @@ def run_migrations_online():
     """
     connectable = engine_from_config(
         config.get_section(config.config_ini_section),
-        prefix=u'sqlalchemy.',
-        poolclass=pool.NullPool)
+        prefix="sqlalchemy.",
+        poolclass=pool.NullPool,
+    )
 
     with connectable.connect() as connection:
         context.configure(
             connection=connection,
             target_metadata=target_metadata,
-            version_table=u'{}_alembic_version'.format(name)
+            version_table="{}_alembic_version".format(name),
         )
 
         with context.begin_transaction():
