+++ ckanext/datavic_reporting/migration/datavic_reporting/versions/1808d0f56068_create_report_job_table.py
@@ -5,13 +5,13 @@ Revises: 390031f5cbb4
 Create Date: 2022-07-05 18:05:16.614061
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 from sqlalchemy.engine.reflection import Inspector
 
 # revision identifiers, used by Alembic.
-revision = '1808d0f56068'
-down_revision = '390031f5cbb4'
+revision = "1808d0f56068"
+down_revision = "390031f5cbb4"
 branch_labels = None
 depends_on = None
 
@@ -27,7 +27,11 @@ def upgrade():
         "report_job",
         sa.Column("id", sa.UnicodeText, primary_key=True),
         sa.Column("report_schedule_id", sa.UnicodeText),
-        sa.Column("timestamp", sa.DateTime, server_default=sa.func.current_timestamp(),),
+        sa.Column(
+            "timestamp",
+            sa.DateTime,
+            server_default=sa.func.current_timestamp(),
+        ),
         sa.Column("filename", sa.UnicodeText),
         sa.Column("frequency", sa.UnicodeText),
         sa.Column("user_roles", sa.UnicodeText),
@@ -36,6 +40,5 @@ def upgrade():
     )
 
 
-
 def downgrade():
-    op.drop_table(    "report_job")
+    op.drop_table("report_job")
