+++ ckanext/datavic_reporting/views/__init__.py
@@ -1,8 +1,7 @@
-from .scheduling import scheduling
-from .reporting import reporting
 from .member_report import member_report
+from .reporting import reporting
+from .scheduling import scheduling
+
 
 def get_blueprints():
-    return [
-        scheduling, reporting, member_report
-    ]
+    return [scheduling, reporting, member_report]
